+++ tilings/strategies/__init__.py
@@ -16,7 +16,7 @@ from .obstruction_inferral import (
     ObstructionTransitivityFactory,
     SubobstructionInferralFactory,
 )
-from .point_jumping import PointJumpingFactory
+from .point_jumping import AssumptionAndPointJumpingFactory
 from .rearrange_assumption import RearrangeAssumptionFactory
 from .requirement_insertion import (
     CellInsertionFactory,
@@ -76,8 +76,9 @@ __all__ = [
     # Equivalence
     "MonotoneSlidingFactory",
     "PatternPlacementFactory",
-    "PointJumpingFactory",
     "SlidingFactory",
+    # Experimental
+    "AssumptionAndPointJumpingFactory",
     # Fusion
     "ComponentFusionFactory",
     "FusionFactory",